// src/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]
// ...
export interface Database {
  public: {
    Tables: {
// ...
      saved_views: {
        Row: {
          id: string
          organization_id: string
          module: string
          name: string
          is_shared: boolean
          is_default: boolean
          created_by_profile_id: string
          state: Json
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string
          organization_id: string
          module: string
          name: string
          is_shared?: boolean
          is_default?: boolean
          created_by_profile_id: string
          state?: Json
          created_at?: string
          updated_at?: string
        }
        Update: {
          id?: string
          organization_id?: string
          module?: string
          name?: string
          is_shared?: boolean
          is_default?: boolean
          created_by_profile_id?: string
          state?: Json
          created_at?: string
          updated_at?: string
        }
      }
    }
// ...
  }
}